third_party/WebKit/LayoutTests/fast/events/constructors/midi-message-event-constructor-expected.txt

This tests the constructor for the MIDIMessageEvent DOM class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new MIDIMessageEvent('eventType').bubbles is false
PASS new MIDIMessageEvent('eventType').cancelable is false
PASS new MIDIMessageEvent('eventType').receivedTime is 0.0
PASS new MIDIMessageEvent('eventType').data is null
PASS new MIDIMessageEvent('eventType', { bubbles: false }).bubbles is false
PASS new MIDIMessageEvent('eventType', { bubbles: true }).bubbles is true
PASS new MIDIMessageEvent('eventType', { cancelable: false }).cancelable is false
PASS new MIDIMessageEvent('eventType', { cancelable: true }).cancelable is true
PASS new MIDIMessageEvent('eventType', { receivedTime: 3.14 }).receivedTime is 3.14
PASS new MIDIMessageEvent('eventType', { data: new Uint8Array(16) }).data.toString() is "[object Uint8Array]"
PASS new MIDIMessageEvent('eventType', { bubbles: true, cancelable: true, receivedTime: 1976.0501, data: new Uint8Array(3) }).bubbles is true
PASS new MIDIMessageEvent('eventType', { bubbles: true, cancelable: true, receivedTime: 1976.0501, data: new Uint8Array(3) }).cancelable is true
PASS new MIDIMessageEvent('eventType', { bubbles: true, cancelable: true, receivedTime: 1976.0501, data: new Uint8Array(3) }).receivedTime is 1976.0501
PASS new MIDIMessageEvent('eventType', { bubbles: true, cancelable: true, receivedTime: 1976.0501, data: new Uint8Array(3) }).data.toString() is "[object Uint8Array]"
PASS successfullyParsed is true

TEST COMPLETE
